refactor: settle the card editor's state before it renders, and drop dead code
Four tidy-ups left over from the review. No behaviour changes.

The card editor worked out which templates exist, and the form to offer for
them, in the middle of render - assigning reactive state as it went, which asks
Lit to schedule another update from inside the one already running. That
preparation now happens in willUpdate, where state is meant to settle, and
render only reads.

Two things were being carried for nothing. `current_view` was written onto Home
Assistant's own lovelace object every time the configuration was looked up and
never read by anything here, and clearTemplateCache was exported but called from
nowhere.

The four repeated stubMember calls that swap a template's body when its type
changes are now a table and a loop, which is what they were describing.

The five test suites each carried their own copy of the same twenty-line check
harness. They share one now.

Also puts back a comment that describes the Share tab: it was left stranded
above the diagnostics method when that was inserted between the two.
