+++ test/variables.test.js
@@ -17,20 +17,7 @@ const {
   normaliseVariables,
 } = require('../.test-build/variables.js');
 
-let passed = 0;
-let failed = 0;
-
-function check(name, actual, expected) {
-  const a = JSON.stringify(actual);
-  const e = JSON.stringify(expected);
-  if (a === e) {
-    passed += 1;
-    console.log(`PASS ${name}`);
-  } else {
-    failed += 1;
-    console.log(`FAIL ${name}\n  got      ${a}\n  expected ${e}`);
-  }
-}
+const { check, report } = require('./harness');
 
 /* ------------------------------------------------------------------ declarations */
 
@@ -311,5 +298,4 @@ check(
 
 check('names of nothing is nothing', forEachNames(undefined), []);
 
-console.log(`\n${passed} passed, ${failed} failed`);
-process.exit(failed ? 1 : 0);
+report();
